feat: Home, and formBooks basic functions finished

+++ src/components/Home.js
@@ -2,6 +2,8 @@ import '../App.css';
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { format } from 'date-fns';
+import Swal from 'sweetalert2';
+import { Button } from "@material-tailwind/react";
 
 function Home() {
   const [list, setList] = useState([])
@@ -16,40 +18,58 @@ function Home() {
   }, [])
 
   const deleteBookCard = async (bid) => {
+    console.log(bid);
+    Swal.fire({
+      title: '¿Estás seguro de borrar esta tarjeta?',
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: 'Si',
+      denyButtonText: `No`,
+    }).then( async (result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        await axios.delete(`http://localhost:8080/api/booksManager/` + bid)
+        const getData = await axios.get("http://localhost:8080/api/booksManager")
+        const data = getData.data.payload.booksCard
+        setList(data)
+        Swal.fire('Borrado!', 'La carta se borró con éxito', 'success')
+      } else if (result.isDenied) {
+        Swal.fire('Cancelado', 'La carta no fue borrada', 'info')
+      }
+    })
     
-    await axios.delete(`http://localhost:8080/api/booksManager/` + bid)
-    const getData = await axios.get("http://localhost:8080/api/booksManager")
-    const data = getData.data.payload.booksCard
-    setList(data)
     
   }
   
   return (
-    <div className=''>
+    <div>
         <h1 className='text-center text-2xl p-10'>Retiro de libros</h1>
-        <div className="flex flex-wrap -m-4 justify-center">
+        <div className="flex flex-wrap justify-center">
+        <div className="choose">
+        <a href="#list-th"><i class="bi bi-grid-fill"></i></a>
+        <a href="#large-th"><i class="fa fa-th-large" aria-hidden="true">b</i></a>
+      </div>
+      </div>
         {list.map((booksCard) => (
-          <div className="p-4 sm:w-1/2 lg:w-1/3">
-            <div className="h-full bg-slate-100 border-2 border-gray-200 border-opacity-60 rounded-lg overflow-hidden">
-              <img className="lg:h-72 md:h-48 w-full object-contain object-center" src={booksCard.books.img} alt="blog"/>
-              <div className="p-6 hover:bg-indigo-700 hover:text-white transition duration-300 ease-in">
-                <h2 className="text-base font-medium text-indigo-300 mb-1"> {booksCard.title} </h2>
-                <h1 className="text-2xl font-semibold mb-3"> {booksCard.books.title} </h1>
-                <p className="leading-relaxed mb-3"> Fecha en la que se retiró: {format(new Date(booksCard.retiredDate), 'dd/MM/yyyy - HH:mm')} </p>
-                <p className="leading-relaxed mb-3"> Fecha de vencimiento: {format(new Date(booksCard.expireDate), 'dd/MM/yyyy - HH:mm')} </p>
-                <div className='flex justify-between'>
-                  <p className="text-indigo-300 inline-flex items-center md:mb-2 lg:mb-0"> {booksCard.owner} </p>
-                  <button className='bg-red-700 p-3 rounded text-white font-bold hover:bg-red-900 transition duration-200 ease-in' onClick={() => deleteBookCard(booksCard._id)}>
-                    Devolver Libro
-                  </button>
-                </div>
-                
-              </div>
-            </div>
+        <div id="list-th">
+        <div className="book read">
+          <div className="cover">
+            <img src={booksCard.books.img} alt='imagen Libro'/>
           </div>
+          <div className="description">
+            <p className='leading-relaxed mb-3 font-bold title'>{booksCard.title}</p>
+            <p className="leading-relaxed mb-3 bookTitle">Libro: {booksCard.books.title}</p>
+            <p className="leading-relaxed mb-3 text-sm horario"> Fecha en la que se retiró: {format(new Date(booksCard.retiredDate), 'dd/MM/yyyy - HH:mm')} </p>
+            <p className="leading-relaxed mb-3 text-sm horario"> Fecha de vencimiento: {format(new Date(booksCard.expireDate), 'dd/MM/yyyy - HH:mm')} </p>
+            <div className="flex justify-center">
+              <Button className='mb-5' color='red' onClick={() => deleteBookCard(booksCard.books._id)}>Borrar</Button>
+            </div>
+            
+          </div>
+        </div>
+        </div>
         ))}
         </div>
-    </div>
   );
 }
